Refactor version check in ChromeDriver VersionResolver
Refactored the way versions are checked in the ChromeDriver's VersionResolver. Replaced array_key_exists() method with a sorted loop and version_compare() for increased accuracy and reliability when comparing versions. This ensures we choose the correct ChromeDriver version for a given browser version, and handles cases where the exact browser version may not be available, finding the closest match instead.